bin/jstor_mapping_demo/requirements.txt
lunr
h_checkmatelib
tldextract